requirements.txt
openai
pillow-avif-plugin
